Clean up plugins for log, updated unit test to use plugins
