+++ src/apps/middleware/logger/CMakeLists.txt
@@ -1,6 +1,5 @@
-add_executable(goby_logger logger.cpp protobuf_logger_plugin.cpp)
+add_executable(goby_logger logger.cpp)
 target_link_libraries(goby_logger goby_middleware)
 
-add_executable(goby_log_tool log_tool.cpp protobuf_logger_plugin.cpp)
+add_executable(goby_log_tool log_tool.cpp)
 target_link_libraries(goby_log_tool goby_middleware)
-
