+++ src/middleware/CMakeLists.txt
@@ -19,7 +19,7 @@ set(SRC
   transport-interprocess-zeromq.cpp
   transport-intervehicle.cpp
   transport-interfaces.cpp
-  log.cpp
+  log/log_entry.cpp
   ${PROTO_SRCS} ${PROTO_HDRS} 
 )
